New idea: only escape tainted strings
Instead of creating a new String subclass like rails, use the taint
mechanism to mark strings as safe or not.

Normally, strings comming from the outside of the ruby process (YAML
config, database values, http requests), should all be tainted. There
are probably some issues with ruby extensions that don't mark the
strings as tainted properly, but it's worth the try.